requirements.txt
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
db-dtypes==1.4.0
exceptiongroup==1.2.2
frozendict==2.4.6
google-api-core==2.24.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html5lib==1.1
idna==3.10
joblib==1.4.2
lxml==5.3.0
multitasking==0.0.11
numpy==1.24.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
pandas-gbq==0.26.1
peewee==3.17.8
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.8.2
PyMuPDF==1.24.11
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
threadpoolctl==3.5.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yfinance==0.2.54
